# Backend_ASP.NET/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Supplies API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Employees/GetByID/{id}": {
      "get": {
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Employees/Update/{id}": {
      "put": {
        "tags": [
          "Employees"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EmployeesModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Employees/Delete/{id}": {
      "delete": {
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Employees/Create": {
      "post": {
        "tags": [
          "Employees"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EmployeesModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Products/GetByID/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Products/Create": {
      "post": {
        "tags": [
          "Products"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProductsModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Stores/GetByID/{id}": {
      "get": {
        "tags": [
          "Stores"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Stores/Create": {
      "post": {
        "tags": [
          "Stores"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StoreModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "EmployeesModel": {
      "type": "object",
      "properties": {
        "employeeId": {
          "format": "uuid",
          "type": "string"
        },
        "hireDate": {
          "format": "date-time",
          "type": "string"
        },
        "salary": {
          "format": "int32",
          "type": "integer"
        },
        "position": {
          "type": "string"
        },
        "status": {
          "type": "boolean"
        },
        "userId": {
          "type": "string"
        },
        "storeID": {
          "format": "uuid",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "ProductsModel": {
      "type": "object",
      "properties": {
        "productID": {
          "format": "uuid",
          "type": "string"
        },
        "productName": {
          "type": "string"
        },
        "price": {
          "format": "int32",
          "type": "integer"
        },
        "stockQuantity": {
          "format": "int32",
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "picture": {
          "type": "string"
        },
        "discontinued": {
          "type": "boolean"
        },
        "storeID": {
          "format": "uuid",
          "type": "string"
        },
        "categoryID": {
          "format": "uuid",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "StoreModel": {
      "type": "object",
      "properties": {
        "storeID": {
          "format": "uuid",
          "type": "string"
        },
        "storeName": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "imageStore": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "Bearer": { }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}